ch(store): modularize store
