+++ app/index.js
@@ -1,37 +1,30 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { applyMiddleware, createStore, combineReducers, compose } from 'redux'
 import { Provider } from 'react-redux'
+import { AppContainer } from 'react-hot-loader'
 
-import thunk from 'redux-thunk'
-import { AppContainer } from 'react-hot-loader';
-
-// local imports
-import users from './redux/reducers/users.js'
+import { configureStore } from './store'
 import getRoutes from './components/App'
-import MainContainer from './components/containers/MainContainer'
 
-const rootReducer = combineReducers({users})
-
-const store = createStore(rootReducer, compose(
-  applyMiddleware(thunk),
-  window.devToolsExtension() ? window.devToolsExtension() : (f) => f
-))
+const store = configureStore()
 
 const renderApp = (Component) => {
   ReactDOM.render(
-    <Provider store={store}>
-      {Component}
-    </Provider>,
+    <AppContainer>
+      <Provider store={store}>
+        <Component />
+      </Provider>
+    </AppContainer>,
     document.getElementById('root')
   )
 }
 
-renderApp(getRoutes())
+// render App first
+renderApp(getRoutes)
 
 if (process.env.NODE_ENV !== 'production' && module.hot) {
   module.hot.accept('./components/App.js', () => {
-    const HotApp = require('./components/App.js').default
+    const HotApp = require('./components/App').default
     // render App on change
     renderApp(HotApp)
   })
