--- app/redux/reducers/index.js
@@ -1,1 +1,0 @@
-export users from './users'
