+++ app/store/index.js
@@ -1,0 +1,1 @@
+export { configureStore } from './store'
