+++ app/store/modules/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux'
+
+import { users } from './users'
+
+export default combineReducers({
+  users,
+})
