+++ app/store/modules/users/index.js
@@ -1,0 +1,1 @@
+export { users } from './users'
